backend/requirements.txt
alembic==1.7.5
aniso8601==9.0.1
black==21.12b0
click==8.0.3
colorama==0.4.4
Flask==2.0.2
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
importlib-metadata==4.8.2
importlib-resources==5.4.0
isort==5.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-sqlalchemy==0.26.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
PyMySQL==1.0.2
python-dotenv==0.19.2
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.27
tomli==1.2.2
typing-extensions==4.0.1
Werkzeug==2.0.2
WTForms==3.0.0
zipp==3.6.0
